Replace BHS with ARE in the list of testing countries.
Having removed BHS, we really needed to replace it in the testing
data.  Additionally, BHS was the only high cement production country
in the testing dataset; all of the others are in the training set.
Emirates has the highest cement production of anyone, so putting it in
the testing set should serve as a kind of a stress test for the
modeling procedure.  Of course, once we start doing actual n-way cross
validation, this won't be such a problem.