Add function predict.simul to simulate forward prediction on retrospective data
In the retrospective data we use the lagged output to add a an
autoregressive term to the model.  When you're using retrospective
data, the lagged output is the actual output, but when you are
predicting outputs for a future scenario, the lagged output will be
the modeled output.  Therefore, to truly evaluate the model
performance, we need to evaluate the retrospective testing data by
predicting outputs from earlier time periods using the model and using
those outputs as the lagged predictors in later time periods.