Compute GDP growth rate as a 5-year trailing annualized rate.
The comments in the version of the code in the previous commit
indicate that we were already doing this, but we weren't.  The rates
in that code were only averaged over one year.  The five-year
averaging takes some of the volatility out of the rates.  It's also
more consistent with how we would actually use the model in GCAM, if
we ever get to that point.